7104147: the fix for cr6887286 was not appropriate for backporting
Summary: Memory leak in the list of ImageObservers watching an object. We use recursion to iterate over the list. So once the list gets too big, we overflow the stack. Fix is to execute a prune operation on ever add, so there is no way we can get indefinite growth of stale nodes.
Reviewed-by: prr, flar