6651256: jstack: DeleteGlobalRef method call doesn't lead to descreasing of global refs count shown by jstack
Summary: jni_DeleteGlobalRef does not really release the jni handle, instead, set the handle point to JNIHandles::_deleted_handle which holds an oop instance (java/lang/Object) in Java heap and never be GC'ed. When counting number of global reference, it counts all the handles on the chain list, which includes the already deleted ones.
Reviewed-by: zgu, sla, coleenp
Contributed-by: yumin.qi@oracle.com